+++ src-docs/src/views/icon/icon_example.js
@@ -19,6 +19,10 @@ import Apps from './apps';
 const appsSource = require('!!raw-loader!./apps');
 const appsHtml = renderToHtml(Apps);
 
+import Ml from './ml';
+const mlSource = require('!!raw-loader!./ml');
+const mlHtml = renderToHtml(Ml);
+
 import Logos from './logos';
 const logosSource = require('!!raw-loader!./logos');
 const logosHtml = renderToHtml(Logos);
@@ -50,7 +54,7 @@ export const IconExample = {
       <div>
         <p>
           <EuiCode>EuiIcon</EuiCode> can build out an icon from our SVG
-          icon library. Icons can be resized and recolored (through a
+          icon library. Icons are resized and recolored (through a
           CSS <EuiCode>Fill</EuiCode>) decleration.
         </p>
         <p>
@@ -86,6 +90,22 @@ export const IconExample = {
       </p>
     ),
     demo: <Apps />,
+  }, {
+    title: 'Machine learning icons',
+    source: [{
+      type: GuideSectionTypes.JS,
+      code: mlSource,
+    }, {
+      type: GuideSectionTypes.HTML,
+      code: mlHtml,
+    }],
+    text: (
+      <p>
+        ML has some specific icons for job creation.
+        Again, these are made for <EuiCode>32x32</EuiCode>.
+      </p>
+    ),
+    demo: <Ml />,
   }, {
     title: 'Logos',
     source: [{
@@ -151,4 +171,3 @@ export const IconExample = {
     demo: <Accessibility />,
   }],
 };
-